PartIterator: int changed to long.

+++ src/main/java/com/sysgears/filesplitter/model/partiterator/IPartIterator.java
@@ -17,5 +17,5 @@ public interface IPartIterator {
      *
      * @return next part size
      */
-    public int nextPartSize();
+    public long nextPartSize();
 }
